embeddings/requirements.txt
bson
flask
python-dotenv
redis
sentence-transformers
waitress
qdrant-client
